model/requirements.txt
requests
flask
flask_cors
pillow
matplotlib
scikit-learn
numpy
pandas
rasterio